requirements.txt
pandas>=2.0.0
numpy>=1.24.0
google-generativeai>=0.3.0
loguru>=0.7.0
playwright>=1.40.0
python-dotenv>=1.0.0
